fix iphop DM
the downloaded data contained 1 folder to much.

`iphop download` creates tgz, md5 and folder (the result of extracting
the tgz) in the folder specified by --dbdir. DB the name of these 3
depends on the reference dataset.

we actually need the inner folder (and also can discard the tgz and md5).

this also adds new reference data